Clean up for migration
